requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.0
graphql-core==3.2.3
h11==0.14.0
idna==3.6
libcst==1.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
strawberry-graphql==0.222.0
typer==0.11.0
typing-inspect==0.9.0
typing_extensions==4.10.0
uvicorn==0.29.0
